+++ src/compas_fab/backends/ros/backend_features/move_it_plan_motion.py
@@ -114,6 +114,7 @@ class MoveItPlanMotion(PlanMotion):
             position=start_configuration.joint_values)
         start_state = RobotState(
             joint_state, MultiDOFJointState(header=header), is_diff=True)
+        start_state.filter_fields_for_distro(self.ros_client.ros_distro)
 
         if options.get('attached_collision_meshes'):
             for acm in options['attached_collision_meshes']: