+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Unreleased
 **Fixed**
 
 * Fixed DH params for analytical IK solver of UR3e and UR10e.
+* Fixed Kinetic support on IK, FK, and motion planning calls.
 
 **Deprecated**